Move wirters to dedicated files
