+++ src/metatrain/utils/data/writers/__init__.py
@@ -1,11 +1,8 @@
-# from pathlib import Path
 from typing import Type
 
-# from metatensor.torch import TensorMap
-# from metatomic.torch import ModelCapabilities, System
-# from .metatensor import write_mts
-# from .xyz import write_xyz
-from .writer import ASEWriter, DiskDatasetWriter, MetatensorWriter, Writer
+from .ase import ASEWriter
+from .metatensor import MetatensorWriter
+from .writers import DiskDatasetWriter, Writer
 
 
 PREDICTIONS_WRITERS = {
@@ -16,12 +13,3 @@ PREDICTIONS_WRITERS = {
 """:py:class:`dict`: dictionary mapping file suffixes to a prediction writers"""
 
 DEFAULT_WRITER: Type[Writer] = ASEWriter
-
-__all__ = [
-    "ASEWriter",
-    "DiskDatasetWriter",
-    "MetatensorWriter",
-    "Writer",
-    "PREDICTIONS_WRITERS",
-    "DEFAULT_WRITER",
-]
